+++ src/basic/entities/coupon/hooks/useCouponForm.ts
@@ -2,12 +2,12 @@ import { useState } from 'react';
 import { CouponType } from '@/types';
 import { initialCouponFormData } from '../constants/coupon';
 
-export const useCouponForm = () => {
+export const useCouponForm = (addCoupon: (coupon: CouponType) => void) => {
   const [showCouponForm, setShowCouponForm] = useState(false);
   const [couponForm, setCouponForm] = useState(initialCouponFormData);
 
   // 쿠폰 폼 제출 처리
-  const handleCouponSubmit = (e: React.FormEvent, addCoupon: (coupon: CouponType) => void) => {
+  const handleCouponSubmit = (e: React.FormEvent) => {
     e.preventDefault();
     addCoupon(couponForm);
     // 폼 초기화
